requirements.txt
genanki==0.13.0
flask==2.3.3
beautifulsoup4==4.12.2
markdown==3.4.4
python-frontmatter==1.0.0
flask-wtf==1.1.1
html2markdown==0.1.7
requests==2.31.0
tqdm==4.66.1 